+++ src/props/rughai/r_tree_s.py
@@ -8,8 +8,8 @@ from engine.utils import animation_set_anchor
 class RTreeS(PositionNode):
     def __init__(
         self,
-        x: int = 0,
-        y: int = 0,
+        x: float = 0,
+        y: float = 0,
         z: float = 0,
         scaling: int = 1,
         batch: Optional[pyglet.graphics.Batch] = None
